git-commit-agent/requirements.txt
# Git Commit Agent Requirements
python-dotenv>=1.0.0
# Add any additional requirements below
